db: fix integer mismatch for download/upload
upload/download are INTEGER in the database.
But while LocationStats (for example) has upload/download defined as
i64, the model has it only as i32.

And when having >i32 values in the db, the log gets flooded with the
following error:
[2026-07-20][10:12:02.440314676][ERROR][webview:vn@tauri://localhost/assets/index-D54brCyI.js:13:34411] Invoking "all_connections" failed due to unknown error: "Database error: error occurred while decoding column 5: out of range integral type conversion attempted"

This change should fix it.

+++ src-tauri/src/lib.rs
@@ -128,6 +128,6 @@ pub struct CommonConnectionInfo {
     pub location_id: Id,
     pub start: NaiveDateTime,
     pub end: NaiveDateTime,
-    pub upload: Option<i32>,
-    pub download: Option<i32>,
+    pub upload: Option<i64>,
+    pub download: Option<i64>,
 }
